properly close websocket session before stopping the server, otherwise error listener is triggered

+++ test/ringo/websocket_worker.js
@@ -38,14 +38,19 @@ const onmessage = function(event) {
 
     let server = new HttpServer();
     let client = new WebSocketClient();
+    let webSocketSession = null;
 
     const stopAll = () => {
-        server.stop();
-        server.destroy();
-        server = null;
-        client.stop();
-        client.destroy();
-        client = null;
+        if (server !== null) {
+            server.stop();
+            server.destroy();
+            server = null;
+        }
+        if (client !== null) {
+            client.stop();
+            client.destroy();
+            client = null;
+        }
     };
 
     const lifeCycleListener = new AbstractLifeCycleListener({
@@ -66,6 +71,7 @@ const onmessage = function(event) {
                     }
                 });
                 listener.on("connect", (session) => {
+                    webSocketSession = session;
                     try {
                         const remote = session.getRemote();
                         if (isBinary) {
@@ -84,6 +90,7 @@ const onmessage = function(event) {
                     } catch (e) {
                         source.postError(e);
                     } finally {
+                        webSocketSession.close();
                         stopAll();
                     }
                 });
@@ -93,6 +100,7 @@ const onmessage = function(event) {
                     } catch (e) {
                         source.postError(e);
                     } finally {
+                        webSocketSession.close();
                         stopAll();
                     }
                 });
@@ -107,7 +115,7 @@ const onmessage = function(event) {
             semaphore.signal();
         }
     });
-    server.jetty.addLifeCycleListener(lifeCycleListener);
+    server.jetty.addEventListener(lifeCycleListener);
 
     const appContext = server.serveApplication("/", function(req) {
         req.charset = 'utf8';
@@ -117,9 +125,11 @@ const onmessage = function(event) {
 
     appContext.addWebSocket(path, function(socket, session) {
         socket.on("text", function(message) {
+            debugger
             socket[isAsync ? "sendString" : "sendStringAsync"](message);
         });
         socket.on("binary", function(bytes, offset, length) {
+            debugger
             socket[isAsync ? "sendBinary" : "sendBinaryAsync"](bytes, offset, length);
         });
     });
